backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Elasticsearch
elasticsearch==8.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1
cryptography==41.0.7  # For Fernet encryption of IMAP passwords

# Email Processing
aiosmtplib==3.0.1
email-validator==2.1.0

# Email Security Analysis
dkimpy==1.1.5
pyspf==2.0.14
py3dns==3.2.1
dnspython==2.4.2
flanker==0.9.11

# Attachment & URL Analysis
python-magic==0.4.27         # File type detection (works cross-platform)
oletools==0.60.1             # Office macro analysis
python-whois==0.8.0          # WHOIS domain age lookup
beautifulsoup4==4.12.2       # HTML parsing
lxml==4.9.3                  # XML/HTML parser
idna==3.6                    # IDN domain encoding/decoding

# SMTP listener
aiosmtpd==1.4.4.post2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
google-generativeai

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
